requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
httpx==0.24.1
